Tested "Ask" loops versus "For" loops. "Ask" loops actually seem faster.
git-svn-id: https://scrap.dis.anl.gov/repos/MRSA/models/agentmodel@220 eb7698ad-c3c5-4393-a60c-5d690fe7bf3d

+++ MRSA/src/mrsa/relogo/Timer.java
@@ -50,9 +50,9 @@ public class Timer {
 		
 		// Report the time.
 		if (counting) {
-			System.out.println(System.nanoTime() - startTime);
+			System.out.println((double) (System.nanoTime() - startTime));
 		} else {
-			System.out.println(stopTime - startTime);
+			System.out.println((double) (stopTime - startTime));
 		}
 		
 	}
